# parallel/pipelined extraction