--- a/Cargo.toml
+++ b/Cargo.toml
@@ -52,6 +52,9 @@ lzma-rs = { version = "0.3.0", default-features = false, optional = true }
 [target.'cfg(any(all(target_arch = "arm", target_pointer_width = "32"), target_arch = "mips", target_arch = "powerpc"))'.dependencies]
 crossbeam-utils = "0.8.20"
 
+[target.'cfg(unix)'.dependencies]
+libc = { version = "0.2.155", optional = true }
+
 [target.'cfg(fuzzing)'.dependencies]
 arbitrary = { version = "1.3.2", features = ["derive"] }
 
@@ -63,6 +66,7 @@ time = { workspace = true, features = ["formatting", "macros"] }
 anyhow = "1"
 clap = { version = "=4.4.18", features = ["derive"] }
 tempdir = "0.3.7"
+tempfile = "3.10.1"
 
 [features]
 aes-crypto = ["aes", "constant_time_eq", "hmac", "pbkdf2", "sha1", "rand", "zeroize"]
@@ -79,6 +83,7 @@ deflate-zopfli = ["zopfli", "_deflate-any"]
 lzma = ["lzma-rs/stream"]
 unreserved = []
 xz = ["lzma-rs/raw_decoder"]
+parallelism = ["libc"]
 default = [
     "aes-crypto",
     "bzip2",
@@ -100,4 +105,8 @@ harness = false
 
 [[bench]]
 name = "merge_archive"
+harness = false
+
+[[bench]]
+name = "extract"
 harness = false