# Problem

[`ZipArchive::extract()`](https://docs.rs/zip/latest/zip/read/struct.ZipArchive.html#method.extract) corresponds to the way most zip implementations perform the task, but it's single-threaded. This is appropriate under the assumptions imposed by rust's `Read` and `Seek` traits, where mutable access is necessary and only one reader can extract file contents at a time, but most unix-like operating systems offer a [`pread()`](https://www.man7.org/linux/man-pages/man2/pread.2.html) operation which avoids mutating OS state like the file offset, so multiple threads can read from a file handle at once. The go programming language offers [`io.ReaderAt`](https://pkg.go.dev/io#ReaderAt) in the stdlib to codify this ability.

# Solution

This is a rework of #72 which avoids introducing unnecessary thread pools and creates all output file handles and containing directories up front. For large zips, we want to:
- create output handles and containing directories up front,
- split the input file handle into chunks to process the constituent file entries in parallel,
- for large compressed entries, pipe their content into a dedicated stream to avoid intermixing i/o and decompression and blocking quick small entries later in the file.

`src/read/split.rs` was created to cover `pread()` and other operations, while `src/read/pipelining.rs` was created to perform the high-level logic to split up entries and perform pipelined extraction.

# Result

- The `parallelism` feature was added to the crate to gate the newly added code + API.
- A dependency on the `libc` crate was added for `#[cfg(all(unix, feature = "parallelism"))]` in order to make use of OS-specific functionality.
- `zip::read::split_extract()` was added as a new external API to extract `&ZipArchive<fs::File>` when `#[cfg(all(unix, feature = "parallelism"))]`.

Note that this does not handle symlinks yet, which I plan to add in a followup PR.

## CURRENT BENCHMARK STATUS

On a linux host (with `splice()` and optionally `copy_file_range()`), we get about a 6.5x speedup with 12 decompression threads:
```shell
> cargo bench --features parallelism -- extract
running 2 tests
test extract_basic           ... bench: 104,389,978 ns/iter (+/- 5,715,453) = 85 MB/s
test extract_split           ... bench:  16,274,974 ns/iter (+/- 1,530,257) = 546 MB/s
```

The performance should keep increasing as we increase thread count, up to the number of available CPU cores (this was running with a parallelism of 12 on my 16-core laptop). This also works on macOS and BSDs, and other `#[cfg(unix)]` platforms.