@cosmicexplorer 2024-07-14
@Pr0methean this is obviously quite a large change, but it's completely separate from any existing APIs and should hopefully remain that way. I was able to avoid adding any new dependencies except for `libc` to make use of `pread()` and `pipe()`, and the external API changes are all gated behind `#[cfg(feature = "parallelism")]`. Please let me know if there are further types of benchmarks or test cases you would like me to add before taking a further look at it, or if you'd prefer to put it under `crate::read::unstable` for now to encourage experimentation. I'm quite enthusiastic that this seems to work for any `#[cfg(unix)]` platform and can be applied to a `&ZipArchive<fs::File>` instead of requiring users to go through any really complex new API.