@cosmicexplorer 2024-07-16
@sluongng thanks so much! I asked for your input here especially so it might be useful in determining the most appropriate way to introduce functionality like prefix trimming. I have introduced a struct `ExtractionParameters` here to the added `split_extract()` method: https://github.com/zip-rs/zip2/blob/bf43e536d36d76b57edef887b9430a5d2a3ac9b2/src/read/pipelining.rs#L700-L738

It's used for perf knobs right now, but could probably be extended to support other configuration, still without messing with the more stable API of the existing `zip` crate.

Regarding `tar` archives: I have been tossing around with @jonjohnsonjr a sort of higher-level library for manipulating zip and tar archives using some of the indexing tools he's been developing for tars/tgzs (https://github.com/jonjohnsonjr/targz), which make them seekable/splittable like we can do with zip files in this PR. I suspect @Pr0methean is interested in keeping the `zip` crate focused on zip files, so in that case what might make sense is for me to take everything @Pr0methean isn't interested in maintaining and pulling it out into that separate library for your purposes.

While I was able to avoid introducing too many extra dependencies (zero except for `libc`) to achieve split extraction in this PR, I am also totally ready for @Pr0methean to say this is too much new code to feel comfortable adding. This is especially reasonable since one of his goals for this crate is safety and support for more features of the zip format, whereas I'm more interested in finding a common subset of zips and tars which allow us to do high-performance splitting and merging (see #193). If so, I would then work to figure out the minimal set of APIs we would want to expose from the `zip` crate to enable this sort of parallelized work in a separate crate. I created [`medusa-zip`](https://github.com/cosmicexplorer/medusa-zip) for this purpose previously when the `zip` crate was unmaintained. and while I've been able to make a lot of changes to the `zip` crate itself already, this might be the point where we find it's time to break off.

@Pr0methean is a fantastic maintainer and is super focused on making this crate very robust, so I'm saying this out loud so we can figure out where to split responsibility for maintenance so nobody gets overwhelmed.